tests: Move test_swf_with_hooks to util/runner
